VNC & Microphone Linux Warning
Warning for VNC and Microphone access if the Control Center is on a
Linux box.

TODO:
VNC: Check if GUI is present for Linux. If so, open reverse connection.
Find Linux software [open source ideal] that will achieve this.
Microphone: Port over richardkiss/speakerpipe utils to Linux, or
repurpose it to output a wav / audio file instead of a raw stream.