Token generation upon user pass phish
Tokens will now be generated even if keychaindump fails, upon a
successful phishing of the user’s password. It uses the user’s password
to decrypt the keychain, and extract the contents of both chrome and
iCloud keys. It then uses these keys to decrypt the databases for each
respective service.